Issue page AlarmReceiver Refactor code and bug start repeat
